add labtest ui, api labtest

+++ clinic/src/main/java/com/hb/service/impl/LabTestsServiceImpl.java
@@ -43,7 +43,7 @@ public class LabTestsServiceImpl implements LabTestService {
         LabTests t = new LabTests();
         String id = params.get("id");
         if (id != null && !id.isEmpty()) {
-            t.setId(Integer.parseInt(id));
+            t.setId(Long.parseLong(id));
         }
 
         t.setTestName(params.get("test_name"));
